+++ monix-execution/jvm/src/test/scala/monix/execution/schedulers/ScheduleOnceJVMSuite.scala
@@ -89,7 +89,7 @@ object ScheduleOnceJVMSuite extends SimpleTestSuite {
   def runTest(sc: Scheduler, threadPrefix: Option[String] = None): Unit = {
     def runAndGetThread(sc: Scheduler, delayMs: Int): Future[String] = {
       val p = Promise[String]()
-      sc.scheduleOnce(delayMs, MILLISECONDS, new Runnable {
+      sc.scheduleOnce(delayMs.toLong, MILLISECONDS, new Runnable {
         def run(): Unit = {
           p.success(Thread.currentThread().getName)
           ()
